chore: Move DiscoveryView to core UI
