--- Mail/Components/Custom Buttons/ModalButtonsView.swift
@@ -1,71 +1,0 @@
-/*
- Infomaniak Mail - iOS App
- Copyright (C) 2024 Infomaniak Network SA
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
-
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-import DesignSystem
-import InfomaniakCoreSwiftUI
-import MailCore
-import MailResources
-import OSLog
-import SwiftUI
-
-struct ModalButtonsView: View {
-    @Environment(\.dismiss) private var dismiss
-
-    @State private var isButtonLoading = false
-
-    let primaryButtonTitle: String
-    var secondaryButtonTitle: String? = MailResourcesStrings.Localizable.buttonCancel
-    var primaryButtonEnabled = true
-    let primaryButtonAction: () async throws -> Void
-    var secondaryButtonAction: (() -> Void)?
-
-    var body: some View {
-        HStack(spacing: IKPadding.micro) {
-            if let secondaryButtonTitle {
-                Button(secondaryButtonTitle) {
-                    secondaryButtonAction?()
-                    dismiss()
-                }
-                .buttonStyle(.ikBorderless)
-                .keyboardShortcut(.cancelAction)
-            }
-
-            Button(primaryButtonTitle) {
-                Task {
-                    isButtonLoading = true
-                    do {
-                        try await primaryButtonAction()
-                        dismiss()
-                    } catch {
-                        Logger.view.warning("\(error)")
-                    }
-                    isButtonLoading = false
-                }
-            }
-            .buttonStyle(.ikBorderedProminent)
-            .disabled(!primaryButtonEnabled)
-            .ikButtonLoading(isButtonLoading)
-            .keyboardShortcut(.defaultAction)
-        }
-        .frame(maxWidth: .infinity, alignment: .trailing)
-    }
-}
-
-#Preview {
-    ModalButtonsView(primaryButtonTitle: "Save") { /* Preview */ }
-}
