+++ Mail/Views/Alerts/CustomScheduleAlertView.swift
@@ -18,6 +18,7 @@
 
 import DesignSystem
 import InfomaniakCoreCommonUI
+import InfomaniakCoreSwiftUI
 import InfomaniakDI
 import MailCore
 import MailCoreUI
