+++ Mail/Views/Alerts/ModifyMessageScheduleAlertView.swift
@@ -17,6 +17,7 @@
  */
 
 import DesignSystem
+import InfomaniakCoreSwiftUI
 import MailCore
 import MailResources
 import SwiftUI
