+++ Mail/Views/New Message/Attachments/AttachmentsUploadInProgressErrorView.swift
@@ -16,6 +16,7 @@
  along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
+import InfomaniakCoreSwiftUI
 import MailResources
 import SwiftUI
