+++ Mail/Views/Settings/General/SettingsThreadModeView.swift
@@ -17,6 +17,7 @@
  */
 
 import InfomaniakCoreCommonUI
+import InfomaniakCoreSwiftUI
 import InfomaniakDI
 import MailCore
 import MailCoreUI
